% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{humanGBMsampleAC}
\alias{humanGBMsampleAC}
\title{Imported dataset of a human glioblastoma spatial transcriptomics sample 
processed with tLOHImportData.}
\format{
A data frame with 34601 rows and 7 variables:
\describe{
  \item{rsID}{dbSNP rs identifier}
  \item{CLUSTER}{cluster number}
  \item{TOTAL}{total number of counts}
  \item{REF}{counts for the reference allele}
  \item{ALT}{counts for the alternative allele}
  \item{CHR}{chromosome number}
  \item{POS}{genomic position}
}
}
\source{
Craig Lab data repository
}
\usage{
humanGBMsampleAC
}
\description{
A dataset of a human glioblastoma sample containing the allele 
count (AC) information for 9 spatial transcriptomics clusters
}
\examples{
data("humanGBMsampleAC")
}
\keyword{datasets}
