% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{initialStartProbabilities}
\alias{initialStartProbabilities}
\title{Imported dataset of sample start probabilities for hiddenMarkovAnalysis}
\format{
A data frame with 22 rows and 2 variables:
\describe{
  \item{initProb1}{Initial Probability 1}
  \item{initProb2}{Initial Probability 2}
}
}
\source{
Craig Lab data repository
}
\usage{
data("initialStartProbabilities")
}
\description{
A dataset of initial start probabilities to use with the HMM
analysis. Users may create their own dataset using the same format
}
\examples{
data("initialStartProbabilities")
}
\keyword{datasets}
