\name{plotLOH}
\alias{plotLOH}
\title{
Plot the output from the tLOHCalc function
}
\description{
Creates a list of two plots for visualiation of allele
frequency and Bayes factors
}
\arguments{
  \item{inputFile}{
  This is an input dataframe with merged cluster data
}
  \item{sample}{
  Name of sample
  }
}

\value{
Output is a list of two plots. Can be assigned to object and visualized
individually
}

\author{
Michelle Webb
}

\examples{
  \dontrun{plotLOH <- function(df,sample)}
}
