\name{plotLOH}
\alias{plotLOH}
\title{
Plot the output from the tLOH function
}
\description{
Creates a list of three plots for visualiation of allele
frequency and bayes factors
}
\arguments{
  \item{inputFile}{
  This is an input file with merged cluster data
}
  \item{sample}{
  Name of sample
  }
}

\value{
Output is a list of three plots. Can be assigned to object and visualized
individually
}

\author{
Michelle Webb
}

\examples{
plotLOH <- function(inputFile,sample)
}
